use size_t in appropriate places

+++ include/geometry/state.h
@@ -1,25 +1,25 @@
 #pragma once
 
+#include <cstddef>
 #include <utility>
 
 namespace geometry {
 
 struct State3d {
-  int x;
-  int y;
-  int theta_bin;
+  size_t x;
+  size_t y;
+  size_t theta_bin;
 
   State3d();
-  State3d(int x, int y, int theta_bin);
+  State3d(size_t x, size_t y, size_t theta_bin);
   bool operator==(const State3d &other) const;
 };
 
-
 struct State2d {
-  int x, y;
+  size_t x, y;
 
-  State2d(int x, int y);
-  State2d(const std::pair<int, int> &s);
+  State2d(size_t x, size_t y);
+  State2d(const std::pair<size_t, size_t> &s);
   State2d(const State3d &s);
 };
 
@@ -27,4 +27,4 @@ struct StateHash {
   std::size_t operator()(const State3d &k) const;
 };
 
-}; // namespace planner
+}; // namespace geometry
